+++ applications/desktop/src/common/commands/index.ts
@@ -2,6 +2,7 @@ export * from "./by-menu/app";
 export * from "./by-menu/cell";
 export * from "./by-menu/edit";
 export * from "./by-menu/file";
+export * from "./by-menu/help";
 export * from "./by-menu/runtime";
 export * from "./by-menu/view";
 export * from "./by-menu/window";
